requirements.txt
# =============================================================================
# MizanAI Chunking - Python Dependencies
# =============================================================================

# Core dependencies
python-dotenv>=1.0.0          # Environment variable management
tiktoken>=0.5.0               # Token counting
numpy>=1.24.0                 # Numerical operations

# Document processing
pdfplumber>=0.10.0            # PDF extraction
pypdf2>=3.0.0                 # PDF processing
python-docx>=1.1.0            # DOCX processing
markdown>=3.5.0               # Markdown processing

# Machine learning (for semantic chunkers)
scikit-learn>=1.3.0           # K-means clustering

# LLM Providers
google-generativeai>=0.3.0    # Google Gemini
openai>=1.0.0                 # OpenAI (also for vLLM, etc.)
ollama>=0.1.0                 # Ollama
litellm>=1.0.0                # LiteLLM (unified interface)

# Vector Stores
chromadb>=0.4.0               # ChromaDB
supabase>=2.0.0               # Supabase
psycopg2-binary>=2.9.0        # PostgreSQL (pgvector)
qdrant-client>=1.7.0          # Qdrant
weaviate-client>=4.0.0        # Weaviate
pinecone-client>=3.0.0        # Pinecone

# Optional dependencies
tqdm>=4.65.0                  # Progress bars
